chore: another code shuffle
I've decided to split out the editor from the game. It should really be its own repo/package/whatever.
I also want to remove the images from the repo. They are really the property of Nintendo and this is a grey area, so be removing the images from the repo, we keep it away from takedown requests